# fixtures/task-solutions/nullable-patch-codegen/api/openapi.yaml
openapi: 3.0.3
info:
  title: Task Service
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /tasks:
    post:
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
    get:
      operationId: listTasks
      responses:
        '200':
          description: Tasks ordered by createdAt and id in ascending order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '500':
          $ref: '#/components/responses/InternalProblem'
  /tasks/{taskId}:
    parameters:
      - $ref: '#/components/parameters/TaskId'
    get:
      operationId: getTask
      responses:
        '200':
          description: Task found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
    delete:
      operationId: deleteTask
      responses:
        '204':
          description: Task deleted
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
    patch:
      operationId: patchTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationProblem'
        '404':
          $ref: '#/components/responses/NotFoundProblem'
        '500':
          $ref: '#/components/responses/InternalProblem'
components:
  parameters:
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Task:
      type: object
      additionalProperties: false
      required:
        - id
        - title
        - createdAt
        - dueAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Trimmed title containing 1 to 200 Unicode code points.
        createdAt:
          type: string
          format: date-time
          x-go-type: string
          description: UTC RFC3339 timestamp returned with exactly six fractional-second digits.
        dueAt:
          type: string
          format: date-time
          nullable: true
          x-go-type: string
          description: Nullable deadline returned in UTC with exactly six fractional-second digits.
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Input is trimmed before validating 1 to 200 Unicode code points.
    TaskList:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Task'
    PatchTaskRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Input is trimmed before validating 1 to 200 Unicode code points.
        dueAt:
          type: string
          format: date-time
          nullable: true
          description: Omitted preserves the deadline, null clears it, and a timestamp sets it.
    Problem:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
          minLength: 1
  responses:
    ValidationProblem:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:problem:validation
            title: Validation failed
            status: 400
            detail: The request is invalid.
    NotFoundProblem:
      description: Task does not exist
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:problem:not-found
            title: Task not found
            status: 404
            detail: The requested task does not exist.
    InternalProblem:
      description: Unexpected server failure
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: urn:problem:internal
            title: Internal server error
            status: 500
            detail: The server could not complete the request.